% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mmultr}
\alias{mmultr}
\title{Matrix multiplication}
\usage{
mmultr(X, Y)
}
\arguments{
\item{X}{Matrix n x p}

\item{Y}{Matrix p x k}
}
\value{
returning Matrix n x k result of multiplication
}
\description{
fast and efficient matrix multiplication with Rcpp
}
\details{
#' \code{mmultr} returns the sum of all the values present in its arguments.
}
